Sub-second fork/suspend

+++ tasks.h
@@ -38,7 +38,7 @@ extern int tasks_set_connection_option(task_queue, const char *,
 extern void new_input_task(task_queue, const char *, int);
 extern void task_suspend_input(task_queue);
 extern enum error enqueue_forked_task2(activation a, int f_index,
-			       unsigned after_seconds, int vid);
+			       double after_seconds, int vid);
 extern enum error enqueue_suspended_task(vm the_vm, void *data);
 				/* data == &(int after_seconds) */
 extern enum error make_reading_task(vm the_vm, void *data);
